# src/main/resources/static/v3/api-docs/swagger.yaml
openapi: 3.1.1
info:
  title: Dataportal Backend REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 7.3.0
...
servers:
  - url: https://to.be.defined
    variables:
      basePath:
        default: /api/v5
...
paths:
...
  /query/feasibility:
    post:
      tags:
        - feasibility
      summary: Create a feasibility query in the broker
      description: The query will be spawned in the broker and directly be dispatched.
      operationId: runQuery
      parameters:
      requestBody:
        description: Structured query to create and dispatch
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CCDL"
        required: true
      responses:
        201:
          description: Query successfully dispatched
          headers:
            Location:
              description: Path to the result of your newly created query
              schema:
                type: string
        401:
          description: Unauthorized - please login first
        403:
          description: Forbidden - insufficient access rights
        422:
          description: Invalid input
        429:
          description: Too many requests in a given amount of time (configurable)
        500:
          description: Dispatch error
      security:
        - dataportal_auth:
            - user
...
  /query/feasibility/cql:
    post:
      tags:
        - feasibility
      summary: Translate a structured query to cql
      description: Nothing will be persisted in the backend. It simply returns the cql.
      operationId: sq2Cql
      parameters:
      requestBody:
        description: Structured query to translate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CCDL"
        required: true
      responses:
        200:
          description: Query successfully translated
        401:
          description: Unauthorized - please login first
        422:
          description: Invalid input
      security:
        - dataportal_auth:
            - user
...
  /query/validate:
    post:
      tags:
        - dataquery
      summary: Validates a submitted CRTDL or just CCDL? to check for schema violations or invalid termCodes
      description: It could be useful to split this in ccdl and dataquery or at least offer a parameter to determine which should be checked.
      operationId: validateQuery
      requestBody:
        description: Query to validate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CRTDL"
        required: true
      responses:
        200:
          description: Query adheres to json schema. If invalid termCodes are present, they will be in the response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CRTDL"
        400:
          description: Query does not adhere to json schema
        401:
          description: Unauthorized - please login first
...
components:
  schemas:
...
    CRTDL:
      type: object
      properties:
        display:
          type: string
          examples:
            - Example CRTDL
        cohortDefinition:
          type: object
          $ref: "#/components/schemas/CCDL"
        dataExtraction:
          type: object
          $ref: "#/components/schemas/DataExtractionQuery"
    DataExtractionQuery:
      type: object
      properties:
        id:
          type: string
          examples:
            - 123
        attributeGroups:
          type: array
          items:
            type: string
            examples:
              - to be done correctly...for now just a placeholder because it doesn't matter for this conversation
...
    CCDL:
      type: object
      required:
        - version
        - inclusionCriteria
      properties:
        version:
          type: string
          format: uri
          description: The json schema version
          examples:
            - http://to_be_decided.com/draft-1/schema#
        display:
          type: string
          examples:
            - foobar
        inclusionCriteria:
          type: array
          items:
            $ref: "#/components/schemas/CriterionList"
        exclusionCriteria:
          type: array
          items:
            $ref: "#/components/schemas/CriterionList"
...
    TermCode:
      description: The termCode defines a concept based on a coding system (i.e. LOINC). The triplet of code, system and version identify the concept.
      type: object
      required:
        - code
        - system
        - display
      properties:
        code:
          type: string
          examples:
            - 119373006
        system:
          type: string
          examples:
            - http://snomed.info/sct
        version:
          type: string
          examples:
            - http://snomed.info/sct/900000000000207008/version/20210731
        display:
          type: string
          examples:
            - Amniotic fluid specimen (specimen)
    beforeDate:
      type: string
      format: date-time
    afterDate:
      type: string
      format: date-time
    TimeRestriction:
      anyOf:
        - $ref: "#/components/schemas/beforeDate"
        - $ref: "#/components/schemas/afterDate"
    Unit:
      type: object
      required:
        - code
        - display
      properties:
        code:
          type: string
        display:
          type: string
    AttributeFilter:
      type: object
      description: An AttributeFilter requires different properties, depending on the type. Please refer to the JSON Schema for this.
      required:
        - type
      properties:
        attributeCode:
          $ref: "#/components/schemas/TermCode"
    ValueFilter:
      type: object
      description: A ValueFilter requires different properties, depending on the type. Please refer to the JSON Schema for this.
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - concept
            - quantity-comparator
            - quantity-range
        selectedConcepts:
          type: array
          items:
            $ref: "#/components/schemas/TermCode"
        comparator:
          type: string
          enum:
            - eq
            - ue
            - le
            - lt
            - ge
            - gt
        unit:
          $ref: "#/components/schemas/Unit"
        value:
          type: number
          format: double
        minValue:
          type: number
          format: double
        maxValue:
          type: number
          format: double
    Criterion:
      type: object
      required:
        - termCodes
      properties:
        termCodes:
          type: array
          items:
            $ref: "#/components/schemas/TermCode"
        attributeFilters:
          type: array
          items:
            $ref: "#/components/schemas/AttributeFilter"
        valueFilter:
          $ref: "#/components/schemas/ValueFilter"
        timeRestriction:
          $ref: "#/components/schemas/TimeRestriction"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    CriterionList:
      type: array
      items:
        $ref: "#/components/schemas/Criterion"
...
    ValidationIssue:
      type: object
      required:
        - code
        - detail
      properties:
        code:
          type: string
          examples:
            - VAL-20001
        detail:
          type: string
          examples:
            - The combination of context and termcode(s) is not found.
...
  securitySchemes:
    dataportal_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://to.be.defined/auth
          scopes:
            user: Dataportal user role
            admin: Dataportal admin role